requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.2
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
packaging==19.1
Pillow==6.1.0
psycopg2==2.8.3
pylint==2.3.1
pyparsing==2.4.2
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.3
wrapt==1.11.2
